move getTImeSince to separate helper file
